Implement mechanism for global preferences
Global preferences are stored in a database table named
`PREFERENCES`.  Each preference is identified by a textual
`key` and has a textual value.  Format of the textual value
is preference specific.

API for reading and writing preferences, as well as definitions
of all preferences, are provided in `src/db/preferences.js`.
In addition to the `key` and an optional `default` value, each
preference definition can provide functions for serializing and
deserializing the preference value.