Implement routes for fetching purchase history
There are three different paths in the API at which
purchase history information can be queried.  In order
to avoid repeating the same code three times, a factory
for Express router which handles the endpoints is provided
at src/routes/admin/purchase_history.js.  This factory
accepts an optional function, which can define additional
conditions for the returned purchases using the Knex builder
API.  The factory is used to create Routers, which are mounted
at the following paths:

 - /api/v1/admin/purchaseHistory
 - /api/v1/admin/users/:userId/purchaseHistory
 - /api/v1/admin/products/:barcode/purchaseHistory

Improving the separation between the bussines logic and database
logic in this commit shoud be considered in the future.