Implement route for deleting a category
Deleting a category moves all products from the
category to be deleted to the default category.
The default category cannot be deleted, as there
is no place to move the existing products.  The
default category is determined by the
`DEFAULT_PRODUCT_CATEGORY` preference, which
contains the `categoryId` of the default category.